Phase 2.3: Break ui.profile <-> ui.action_model direct cycle
- Created gremlin/ui/models.py for shared UI data structures
- Moved SequenceIndex from action_model.py to models.py
- Both ui.profile and ui.action_model now import from models.py
- Direct circular dependency broken (though indirect cycle remains via plugins)
- Applied Separation of Concerns principle
