+++ gremlin/ui/action_model.py
@@ -30,6 +30,8 @@ from gremlin.signal import signal
 from gremlin.types import ActionActivationMode, InputType
 # REFACTORED: Use IBindingModel abstraction to break circular dependency
 from gremlin.domain import IBindingModel
+# REFACTORED: Import SequenceIndex from shared models module
+from gremlin.ui.models import SequenceIndex
 
 if TYPE_CHECKING:
     from gremlin.base_classes import AbstractActionData
@@ -40,39 +42,7 @@ QML_IMPORT_NAME = "Gremlin.Profile"
 QML_IMPORT_MAJOR_VERSION = 1
 
 
-class SequenceIndex:
-
-    def __init__(
-            self,
-            parent_index: int | None,
-            container_name: str | None,
-            index: int,
-    ):
-        """Creates a new action index instance.
-        This models the QModelIndex class.
-        Args:
-            parent_index: index assigned to the parent action
-            container_name: name of the parent's container
-            index: index assigned to this action
-        """
-        self._parent_index = parent_index
-        self._container_name = container_name
-        self._index = index
-
-    @property
-    def index(self) -> int:
-        return self._index
-
-    @property
-    def parent_index(self) -> int:
-        return self._parent_index
-
-    @property
-    def container_name(self) -> str:
-        return self._container_name
-
-    def __str__(self) -> str:
-        return f"SID: c={self.container_name}: p={self.parent_index} i={self.index}"
+# SequenceIndex moved to gremlin.ui.models to break circular dependency
 
 
 @QtQml.QmlElement
